refactor to atomic directory
